% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_fastqs}
\alias{df_fastqs}
\title{flattened dataframe of 4 samples}
\format{
A data frame with 204 rows and 8 variables:
\describe{
\item{Base}{The base number}
\item{Mean}{The Mean}
\item{Lower Quartile}{The bottom quartile}
\item{Upper Quartile}{The top quartile}
\item{10th Percentile}{Bottom 10\%}
\item{90th Percentile}{Top 10\%}
\item{sample_names}{the character vector}
}
}
\usage{
df_fastqs
}
\description{
A dataset containing the 2 biological replicates of 2 technical replicates each
The variables are as follows:
}
\keyword{datasets}
