% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{get_names}
\alias{get_names}
\title{get names}
\usage{
get_names(file_names, regexp = "WT_rep[12]_ERR[0-9]{6}")
}
\arguments{
\item{file_names}{pass file names, must be a character vector}

\item{regexp}{regexp to look for in file name. Default \verb{WT_rep[12]_ERR[0-9]\{6\}}}
}
\value{
list of character strings
}
\description{
get names
}
